requirements.txt
fastapi
uvicorn[standard]
telnyx
deepgram-sdk
PyNaCl
